feat(workspace): export boringDefaultFrontPlugins from @hachej/boring-workspace/app/vite
Extracts the Vite virtual module plugin into a shared export so all
apps drive their front plugin list from boring.defaultPlugins in their
own package.json — same single source of truth as the server side.

- Add packages/workspace/src/app/vite/index.ts with boringDefaultFrontPlugins({ appRoot })
- Wire as new tsup entry (app-vite) + ./app/vite package.json export
- workspace-playground imports from the shared export instead of defining inline
- CLI vite.config.ts adds the plugin; App.tsx replaces static askUserPlugin
  import with virtual:boring-front-plugins
