+++ packages/cli/src/front/App.tsx
@@ -4,7 +4,7 @@ import * as ReactDomClient from "react-dom/client"
 import * as ReactJsxDevRuntime from "react/jsx-dev-runtime"
 import * as ReactJsxRuntime from "react/jsx-runtime"
 import { useCallback, useEffect, useMemo, useState } from "react"
-import { askUserPlugin } from "@hachej/boring-ask-user/front"
+import defaultFrontPlugins from "virtual:boring-front-plugins"
 import * as WorkspaceSingleton from "@hachej/boring-workspace"
 import * as WorkspaceEventsSingleton from "@hachej/boring-workspace/events"
 import * as WorkspacePluginSingleton from "@hachej/boring-workspace/plugin"
@@ -177,7 +177,7 @@ export function CliWorkspaceShell() {
     setUrlSessionId((current) => current === sessionId ? current : sessionId)
   }, [])
 
-  const plugins = useMemo(() => [askUserPlugin], [])
+  const plugins = useMemo(() => [...defaultFrontPlugins], [])
   const activeWorkspaceRequestHeaders = useMemo(
     () => activeWorkspaceId ? { "x-boring-workspace-id": activeWorkspaceId } : null,
     [activeWorkspaceId],
