+++ packages/cli/src/front/virtual.d.ts
@@ -1,0 +1,5 @@
+declare module "virtual:boring-front-plugins" {
+  import type { BoringFrontPlugin } from "@hachej/boring-workspace"
+  const plugins: BoringFrontPlugin[]
+  export default plugins
+}
